+++ doc/changelog.rst
@@ -1,6 +1,18 @@
 Change Log
 ==========
 
+Version 3.0.1 (2022-05-26)
+----------------------------
+Fixes
+^^^^^
+- Make sure to include _hyperlinks in history items
+- On edit ticket, raise exception if user/queue does not exist
+
+Added
+^^^^^
+- Add helper method for deleting tickets
+- Add tests
+
 Version 3.0.0 (2022-05-17)
 ----------------------------
 The following is a major release of the `rt` library.